Move WhitehallImporter to lib directory
This commit moves the WhitehallImporter to the root of the lib
directory as it didn't really belong under tasks as it was a script
instead of a rake task. Also, as we refactor the code into various
classes it makes more sense to reside here.
