+++ spec/lib/notify_rake_spec.rb
