+++ spec/lib/unpublish_rake_spec.rb
