+++ spec/lib/whitehall_importer_spec.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-RSpec.describe Tasks::WhitehallImporter do
+RSpec.describe WhitehallImporter do
   include FixturesHelper
 
   let(:import_data) { whitehall_export_with_one_edition }
 
   it "can import JSON data from Whitehall" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
 
     expect { importer.import }.to change { Document.count }.by(1)
 
@@ -24,7 +24,7 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "adds users who have never logged into Content Publisher" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(User.last.uid).to eq "36d5154e-d3b7-4e3e-aad8-32a50fc9430e"
@@ -35,14 +35,14 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "does not add users who have logged into Content Publisher" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     User.create!(uid: "36d5154e-d3b7-4e3e-aad8-32a50fc9430e")
 
     expect { importer.import }.not_to(change { User.count })
   end
 
   it "creates a user map" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expected_user_ids = {
@@ -53,14 +53,14 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "sets created_by_id as the original author" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Document.last.created_by_id).to eq(User.last.id)
   end
 
   it "sets import_from as Whitehall" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     document = Document.last
@@ -74,7 +74,7 @@ RSpec.describe Tasks::WhitehallImporter do
       "state" => "published",
       "whodunnit" => 1,
     }
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.status).to be_published
@@ -83,7 +83,7 @@ RSpec.describe Tasks::WhitehallImporter do
 
   it "can set minor update type" do
     import_data["editions"][0]["minor_change"] = true
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.update_type).to eq("minor")
@@ -98,7 +98,7 @@ RSpec.describe Tasks::WhitehallImporter do
       "state" => "published",
       "whodunnit" => 1,
     }
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.status).to be_published_but_needs_2i
@@ -112,7 +112,7 @@ RSpec.describe Tasks::WhitehallImporter do
       "state" => "rejected",
       "whodunnit" => 1,
     }
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.status).to be_submitted_for_review
@@ -126,7 +126,7 @@ RSpec.describe Tasks::WhitehallImporter do
       "state" => "submitted",
       "whodunnit" => 1,
     }
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.status).to be_submitted_for_review
@@ -150,22 +150,22 @@ RSpec.describe Tasks::WhitehallImporter do
 
   it "raises AbortImportError when edition has an unsupported state" do
     import_data["editions"][0]["state"] = "not_supported"
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
 
-    expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+    expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
   end
 
   it "raises AbortImportError when revision history is missing for state" do
     import_data["editions"][0]["state"] = "published"
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
 
-    expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+    expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
   end
 
   it "sets the created_at datetime of the document state" do
     import_data["editions"][0]["revision_history"][0].merge!("created_at" => 3.days.ago)
 
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     imported_created_at = import_data["editions"][0]["revision_history"][0]["created_at"]
@@ -175,16 +175,16 @@ RSpec.describe Tasks::WhitehallImporter do
 
   it "raises AbortImportError when edition has an unsupported locale" do
     import_data["editions"][0]["translations"][0]["locale"] = "zz"
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
 
-    expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+    expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
   end
 
   it "changes the ids of embedded contacts" do
     import_data["editions"][0]["translations"][0]["body"] = "[Contact:123]"
     content_id = SecureRandom.uuid
     import_data["editions"][0]["contacts"] = [{ "id" => 123, "content_id" => content_id }]
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     expect(Edition.last.contents["body"]).to eq("[Contact:#{content_id}]")
@@ -192,7 +192,7 @@ RSpec.describe Tasks::WhitehallImporter do
 
   context "when importing organisation associations" do
     it "sets a primary_publishing_organisation" do
-      importer = Tasks::WhitehallImporter.new(import_data)
+      importer = WhitehallImporter.new(import_data)
       importer.import
 
       imported_organisation = import_data["editions"][0]["organisations"][0]
@@ -203,16 +203,16 @@ RSpec.describe Tasks::WhitehallImporter do
 
     it "rejects the import if there are no organisations" do
       import_data["editions"][0].delete("organisations")
-      importer = Tasks::WhitehallImporter.new(import_data)
+      importer = WhitehallImporter.new(import_data)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
 
     it "rejects the import if there are no lead organisations" do
       import_data["editions"][0]["organisations"].shift
-      importer = Tasks::WhitehallImporter.new(import_data)
+      importer = WhitehallImporter.new(import_data)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
 
     it "rejects the import if there is more than one lead organisation" do
@@ -223,13 +223,13 @@ RSpec.describe Tasks::WhitehallImporter do
         "lead_ordering" => 2,
       )
 
-      importer = Tasks::WhitehallImporter.new(import_data)
+      importer = WhitehallImporter.new(import_data)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
 
     it "sets other supporting organisations" do
-      importer = Tasks::WhitehallImporter.new(import_data)
+      importer = WhitehallImporter.new(import_data)
       importer.import
 
       imported_organisation = import_data["editions"][0]["organisations"][1]
@@ -240,7 +240,7 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "sets role appointments" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     imported_role_appointment = import_data["editions"][0]["role_appointments"][0]
@@ -250,7 +250,7 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "sets topical events" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     imported_topical_events = import_data["editions"][0]["topical_events"][0]
@@ -260,7 +260,7 @@ RSpec.describe Tasks::WhitehallImporter do
   end
 
   it "sets world locations" do
-    importer = Tasks::WhitehallImporter.new(import_data)
+    importer = WhitehallImporter.new(import_data)
     importer.import
 
     imported_world_locations = import_data["editions"][0]["world_locations"][0]
@@ -273,7 +273,7 @@ RSpec.describe Tasks::WhitehallImporter do
     let(:import_published_then_drafted_data) { whitehall_export_with_two_editions }
 
     it "only creates the latest edition" do
-      importer = Tasks::WhitehallImporter.new(import_published_then_drafted_data)
+      importer = WhitehallImporter.new(import_published_then_drafted_data)
       importer.import
 
       expect(Edition.last.status).to be_draft
@@ -281,14 +281,14 @@ RSpec.describe Tasks::WhitehallImporter do
     end
 
     it "sets imported to true on revision" do
-      importer = Tasks::WhitehallImporter.new(import_published_then_drafted_data)
+      importer = WhitehallImporter.new(import_published_then_drafted_data)
       importer.import
 
       expect(Revision.last.imported).to be true
     end
 
     it "sets created_by_id on each edition as the original edition author" do
-      importer = Tasks::WhitehallImporter.new(import_published_then_drafted_data)
+      importer = WhitehallImporter.new(import_published_then_drafted_data)
       importer.import
 
       expect(Edition.second_to_last.created_by_id).to eq(User.second_to_last.id)
@@ -296,7 +296,7 @@ RSpec.describe Tasks::WhitehallImporter do
     end
 
     it "sets last_edited_by_id on each edition as the most recent author" do
-      importer = Tasks::WhitehallImporter.new(import_published_then_drafted_data)
+      importer = WhitehallImporter.new(import_published_then_drafted_data)
       importer.import
 
       expect(Edition.second_to_last.last_edited_by_id).to eq(User.second_to_last.id)
@@ -305,9 +305,9 @@ RSpec.describe Tasks::WhitehallImporter do
 
     it "raises AbortImportError when an edition has an unsupported document type" do
       import_published_then_drafted_data["editions"][0]["news_article_type"] = "unsupported_document"
-      importer = Tasks::WhitehallImporter.new(import_published_then_drafted_data)
+      importer = WhitehallImporter.new(import_published_then_drafted_data)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
   end
 
@@ -315,7 +315,7 @@ RSpec.describe Tasks::WhitehallImporter do
     let(:import_data_for_withdrawn_edition) { whitehall_export_with_one_withdrawn_edition }
 
     it "sets the correct states when Whitehall document state is withdrawn" do
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
       importer.import
 
       expect(Status.count).to eq(2)
@@ -327,7 +327,7 @@ RSpec.describe Tasks::WhitehallImporter do
     it "sets the correct states when Whitehall document state is withdrawn and was force_published" do
       import_data_for_withdrawn_edition["editions"][0]["force_published"] = true
 
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
       importer.import
 
       expect(Status.count).to eq(2)
@@ -337,7 +337,7 @@ RSpec.describe Tasks::WhitehallImporter do
     end
 
     it "sets the created_by_id of each status if more than one state needs to be recorded" do
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
       importer.import
 
       expect(Status.first.created_by_id).to eq(User.second_to_last.id)
@@ -346,13 +346,13 @@ RSpec.describe Tasks::WhitehallImporter do
 
     it "raises AbortImportError when revision history cannot be found for state" do
       import_data_for_withdrawn_edition["editions"][0]["revision_history"].delete_at(1)
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
 
     it "sets the created_at datetime of the initial and current document states" do
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
       importer.import
 
       import_revision_history = import_data_for_withdrawn_edition["editions"][0]["revision_history"]
@@ -363,13 +363,13 @@ RSpec.describe Tasks::WhitehallImporter do
 
     it "raises AbortImportError when document is withdrawn but has no unpublishing details" do
       import_data_for_withdrawn_edition["editions"][0]["unpublishing"] = nil
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
 
-      expect { importer.import }.to raise_error(Tasks::WhitehallImporter::AbortImportError)
+      expect { importer.import }.to raise_error(WhitehallImporter::AbortImportError)
     end
 
     it "sets the Withdrawal details for a withdrawn document" do
-      importer = Tasks::WhitehallImporter.new(import_data_for_withdrawn_edition)
+      importer = WhitehallImporter.new(import_data_for_withdrawn_edition)
       importer.import
 
       import_unpublishing_data = import_data_for_withdrawn_edition["editions"][0]["unpublishing"]
